try1 for booleantransformation
